MalodyBSC-web/requirements.txt
flask
redis
python-redis-lock